official_writeup/txspeech/game/requirements.txt
tencentcloud-sdk-python
flask
uwsgi